Address review feedback: restructure as package, fix docstrings, move changes entry
- Convert _asyncioEventLoop.py to _asyncioEventLoop/ package
- __init__.py exposes only initialize() and terminate()
- utils.py contains runCoroutine() and runCoroutineSync()
- Shared mutable state (eventLoop, asyncioThread) in _state.py
- Use triple-quoted docstrings for module-level variable annotations
- Move changes.md entry to correct position in developer changes list
- Update tests to import from utils module directly
