+++ tests/unit/test_asyncioEventLoop.py
@@ -9,6 +9,7 @@ import asyncio
 import unittest
 
 import _asyncioEventLoop
+from _asyncioEventLoop.utils import runCoroutineSync
 
 
 class TestRunCoroutineSync(unittest.TestCase):
@@ -30,7 +31,7 @@ class TestRunCoroutineSync(unittest.TestCase):
 		async def simpleCoroutine():
 			return 42
 
-		result = _asyncioEventLoop.runCoroutineSync(simpleCoroutine())
+		result = runCoroutineSync(simpleCoroutine())
 		self.assertEqual(result, 42)
 
 	def test_returnsComplexResult(self):
@@ -40,7 +41,7 @@ class TestRunCoroutineSync(unittest.TestCase):
 			await asyncio.sleep(0.01)
 			return {"key": "value", "number": 123}
 
-		result = _asyncioEventLoop.runCoroutineSync(complexCoroutine())
+		result = runCoroutineSync(complexCoroutine())
 		self.assertEqual(result, {"key": "value", "number": 123})
 
 	def test_raisesException(self):
@@ -51,7 +52,7 @@ class TestRunCoroutineSync(unittest.TestCase):
 			raise ValueError("Test error message")
 
 		with self.assertRaises(ValueError) as cm:
-			_asyncioEventLoop.runCoroutineSync(failingCoroutine())
+			runCoroutineSync(failingCoroutine())
 		self.assertEqual(str(cm.exception), "Test error message")
 
 	def test_timeoutRaisesTimeoutError(self):
@@ -62,7 +63,7 @@ class TestRunCoroutineSync(unittest.TestCase):
 			return "Should not reach here"
 
 		with self.assertRaises(TimeoutError) as cm:
-			_asyncioEventLoop.runCoroutineSync(slowCoroutine(), timeout=0.1)
+			runCoroutineSync(slowCoroutine(), timeout=0.1)
 		self.assertIn("timed out", str(cm.exception).lower())
 
 	def test_noTimeoutWaitsIndefinitely(self):
@@ -72,24 +73,23 @@ class TestRunCoroutineSync(unittest.TestCase):
 			await asyncio.sleep(0.1)
 			return "completed"
 
-		# This should complete successfully even though it takes some time
-		result = _asyncioEventLoop.runCoroutineSync(delayedCoroutine())
+		result = runCoroutineSync(delayedCoroutine())
 		self.assertEqual(result, "completed")
 
 	def test_raisesRuntimeErrorWhenEventLoopNotRunning(self):
 		"""Test that runCoroutineSync raises RuntimeError when event loop is not running."""
 		# Save original thread reference
-		originalThread = _asyncioEventLoop.asyncioThread
+		originalThread = _asyncioEventLoop._state.asyncioThread
 
 		# Temporarily set to None to simulate not running
-		_asyncioEventLoop.asyncioThread = None
+		_asyncioEventLoop._state.asyncioThread = None
 
 		async def anyCoroutine():
 			return "test"
 
 		with self.assertRaises(RuntimeError) as cm:
-			_asyncioEventLoop.runCoroutineSync(anyCoroutine())
+			runCoroutineSync(anyCoroutine())
 		self.assertIn("not running", str(cm.exception).lower())
 
 		# Restore original thread
-		_asyncioEventLoop.asyncioThread = originalThread
+		_asyncioEventLoop._state.asyncioThread = originalThread
